# Refactor stream reads