This PR refactors how data is read and uncompressed.

It adds a `doubleBuffer` helper struct to do the magic C, double buffer dance and makes use of this in the `Read` function.

Here are the benchmarks comparing the current `master` with this branch:

#### master

```
$ go test -bench='^BenchmarkStreamUncompress.*' -benchtime 15s -count 2
goos: darwin
goarch: amd64
pkg: github.com/DataDog/golz4
BenchmarkStreamUncompress-4   	 5527802	      4236 ns/op	7735.24 MB/s	     123 B/op	       2 allocs/op
BenchmarkStreamUncompress-4   	 3750264	      6997 ns/op	4682.91 MB/s	     142 B/op	       2 allocs/op
PASS
ok  	github.com/DataDog/golz4	65.629s
```

#### New version

```
$ go test -bench='^BenchmarkStreamUncompress.*' -benchtime 15s -count 2
goos: darwin
goarch: amd64
pkg: github.com/DataDog/golz4
BenchmarkStreamUncompress-4   	 5643277	      4133 ns/op	7927.65 MB/s	     126 B/op	       3 allocs/op
BenchmarkStreamUncompress-4   	 3755970	      6837 ns/op	4792.58 MB/s	     140 B/op	       3 allocs/op
PASS
ok  	github.com/DataDog/golz4	65.232s
```

Performance is very similar, sometimes a bit better. There's still an extra alloc that slipped through the cracks, probably related to the addition of `doubleBuffer`.

FWIW, I also tested moving the  temporary `readBuffer` as a struct member but this turned out to be worse, performance wise