@j-vizcaino 2021-02-08
Looks like this PR got superseded by other recent work on readers. Closing.